hi3

+++ src/main/java/unit2/Unit2.java
@@ -66,7 +66,7 @@ public class Unit2 {
      */
     public static int roundValue(double value) {
         // TODO: Implement this method using Math.round
-        return Math.round(value);
+        return (int)Math.round(value);
     }
 
     /**
